Что то сломал

+++ app/src/main/java/ru/kbs41/kbsdatacollector/retrofit/ExchangeMaster.kt
@@ -99,7 +99,7 @@ class ExchangeMaster {
                 t.sourceGuid,
                 t.rowNumber,
                 t.qty,
-                0.toDouble(),
+                0.0,
                 orderId,
                 productDao.getProductByGuid(t.productSourceId)[0].id
             )
